+++ vault/resource_jwt_auth_backend.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"log"
+	"sort"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
@@ -122,11 +123,40 @@ func jwtAuthBackendResource() *schema.Resource {
 				Description: "The accessor of the JWT auth backend",
 			},
 			"provider_config": {
-				Type:        schema.TypeMap,
+				Type:        schema.TypeSet,
 				Optional:    true,
 				Description: "Provider specific handling configuration",
-				Elem: &schema.Schema{
-					Type: schema.TypeString,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"provider": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+						"gsuite_service_account": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+						"gsuite_admin_impersonate": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+						"fetch_groups": {
+							Type:     schema.TypeBool,
+							Optional: true,
+						},
+						"fetch_user_info": {
+							Type:     schema.TypeBool,
+							Optional: true,
+						},
+						"groups_recurse_max_depth": {
+							Type:     schema.TypeInt,
+							Optional: true,
+						},
+						"user_custom_schemas": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+					},
 				},
 			},
 			"tune": authMountTuneSchema(),
@@ -263,7 +293,25 @@ func jwtAuthBackendRead(d *schema.ResourceData, meta interface{}) error {
 		if configOption == "oidc_client_secret" {
 			continue
 		}
-		d.Set(configOption, config.Data[configOption])
+
+		if configOption == "provider_config" {
+			var keys []string
+			rawProviderConfig := config.Data[configOption].(map[string]interface{})
+			for k := range rawProviderConfig {
+				keys = append(keys, k)
+			}
+
+			// The keys need to be sorted to ensure the hash is calculated correctly.
+			sort.Strings(keys)
+			result := make([]interface{}, len(keys))
+			for i, k := range keys {
+				result[i] = rawProviderConfig[k]
+			}
+
+			d.Set(configOption, result)
+		} else {
+			d.Set(configOption, config.Data[configOption])
+		}
 	}
 
 	log.Printf("[DEBUG] Reading jwt auth tune from %q", path+"/tune")
@@ -289,8 +337,14 @@ func jwtAuthBackendUpdate(d *schema.ResourceData, meta interface{}) error {
 	configuration := map[string]interface{}{}
 	for _, configOption := range matchingJwtMountConfigOptions {
 		// Set the configuration if the user has specified it, or the attribute is in the Diff
-		if _, ok := d.GetOkExists(configOption); ok || d.HasChange(configOption) {
-			configuration[configOption] = d.Get(configOption)
+		if configOption != "provider_config" {
+			if _, ok := d.GetOkExists(configOption); ok || d.HasChange(configOption) {
+				configuration[configOption] = d.Get(configOption)
+			}
+		} else {
+			if _, ok := d.GetOkExists(configOption); ok || d.HasChange(configOption) {
+				configuration[configOption] = d.Get(configOption).(*schema.Set).List()
+			}
 		}
 	}